+++ app/src/components/ParticipantWindow/fragments/FailureBadge.tsx
@@ -3,9 +3,9 @@
 // SPDX-License-Identifier: EUPL-1.2
 import { Stack, Tooltip } from '@mui/material';
 import { WarningIcon } from '@opentalk/common';
-import React from 'react';
+import { ReactNode } from 'react';
 
-export const FailureBadge = ({ title, children }: { title: string; children: React.ReactNode }) => (
+export const FailureBadge = ({ title, children }: { title: string; children: ReactNode }) => (
   <Tooltip title={title}>
     <Stack direction={'row'}>
       <WarningIcon color="error" fontSize="medium" />